Crate futures_tls

Async TLS streams

This library is an implementation of TLS streams using the most appropriate system library by default for negotiating the connection. That is, on Windows this library uses SChannel, on OSX it uses SecureTransport, and on other platforms it uses OpenSSL. The usage of OpenSSL can optionally be forced with the force-openssl feature of this crate.

Each TLS stream implements traits from the futures_io crate to interact and interoperate with the rest of the futures I/O ecosystem. Client connections initiated from this crate verify hostnames automatically and by default.

Modules

backend
Backend-specific extension traits.

Structs

ClientContext
A context used to initiate client-side connections to a TLS server.

ClientHandshake
A future returned from ClientContext::handshake used to represent an in-progress TLS handshake.

ServerContext
A context used to initiate server-side connections of a TLS server.

ServerHandshake
A future returned from ServerContext::handshake used to represent an in-progress TLS handshake.

TlsStream
A wrapper around an underlying raw stream which implements the TLS or SSL protocol.