Struct futures_tls::ClientContext

pub struct ClientContext {
    // some fields omitted
}

A context used to initiate client-side connections to a TLS server.

Client context by default perform hostname validation of connections issued by ensuring that certificates sent by the server are trusted by the system and are indeed valid.

Normally a ClientContext doesn't need extra configuration, but extra configuration can be performed by the backend-specific extension traits.

Methods

impl ClientContext

fn new() -> Result<ClientContext>

Creates a new client context ready for connecting to a remote server.

The client context can be optionally configured through backend-specific extension traits, but by default client contexts will verify certificates against the system certificate store and otherwise verify that certificates are indeed valid.

fn handshake<S>(self, domain: &str, stream: S) -> ClientHandshake<S> where S: Read + Write + Stream<Item=Ready, Error=Error>

Performs a handshake with the given I/O stream to resolve to an actual I/O stream.

This function will consume this context and return a future which will either resolve to a TlsStream<S> ready for reading/writing if the handshake completes successfully, or an error if an erroneous event otherwise happens.

The provided domain argument is the domain that this client intended to connect to. The TLS/SSL backend will verify that the certificate provided by the server indeed matches the domain provided. The returned future will only resolve successfully if the domain matches, otherwise an error will be returned.

The given I/O stream should be a freshly connected client (typically a TCP stream) ready to negotiate the TLS connection.