Struct futures_tls::ServerContext

pub struct ServerContext {
    // some fields omitted
}

A context used to initiate server-side connections of a TLS server.

Server contexts are typically much harder to create than a client context because they need to know the public/private key that they're going to negotiate the connection with. Specifying these keys is typically done in a very backend-specific method, unfortunately. For that reason there's no new constructor.

For some examples of how to create a context, though, you can take a look at the test suite of futures-tls.

Methods

impl ServerContext

fn handshake<S>(self, stream: S) -> ServerHandshake<S> where S: Read + Write + Stream<Item=Ready, Error=Error>

Performs a handshake with the given I/O stream to resolve to an actual I/O stream.

This function will consume this context and return a future which will either resolve to a TlsStream<S> ready for reading/writing if the handshake completes successfully, or an error if an erroneous event otherwise happens.

The given I/O stream should be an accepted client of this server which is ready to negotiate the TLS connection.